wip balance spec: blocked on creating paypal balance
